requirements.txt
torch>=1.13
torchvision
transformers
ultralytics
pandas
numpy
matplotlib
pyserial
